Numerical Integration for Elliptical Segment Area
An ellipse centered at the origin is plotted using two parametric equations. Using f(t) = a*cos(t) and g(t) = b*sin(t), g(f(t)) was plotted in Cartesian coordinates over a domain that passes the vertical line test, (0<= t <= pi). The resulting elliptical segment was curve fitted* and numerically integrated; the area of an isosceles triangular region with points at the origin and boundaries of the domain was subtracted to find the area of the segment*